requirements.txt
Django==2.0.6
django-allauth==0.36.0
django-appconf==1.0.2
django-allauth==0.36.0
django-cors-headers==2.2.0
django-extensions==2.0.7
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Werkzeug==0.14.1
whitenoise==3.3.1
pyOpenSSL==18.0.0
PyJWT==1.6.4
certifi==2018.4.16
